+++ ExtensionMethods/ShortNullable.cs
@@ -78,7 +78,7 @@ public static partial class CheckValidatorsExtensions
     /// <param name="value">The number you are comparing</param>
     /// <param name="msg">Custom error message</param>
     /// <returns></returns>
-    public static Check<short?> IfGreaterThan(this Check<short?> data, int value)
+    public static Check<short?> IfGreaterThan(this Check<short?> data, short value)
     {
         if (data.InvalidModel()) { return data; }
         if (data.Value > value)
@@ -95,7 +95,7 @@ public static partial class CheckValidatorsExtensions
     /// <param name="value">The number you are comparing</param>
     /// <param name="msg">Custom error message</param>
     /// <returns></returns>
-    public static Check<short?> IfLessThan(this Check<short?> data, int value)
+    public static Check<short?> IfLessThan(this Check<short?> data, short value)
     {
         if (data.InvalidModel()) { return data; }
         if (data.Value > value)
@@ -112,7 +112,7 @@ public static partial class CheckValidatorsExtensions
     /// <param name="value">The number you are comparing</param>
     /// <param name="msg">Custom error message</param>
     /// <returns></returns>
-    public static Check<short?> IfEquals(this Check<short?> data, int value)
+    public static Check<short?> IfEquals(this Check<short?> data, short value)
     {
         if (data.InvalidModel()) { return data; }
         if (data.Value == value)
@@ -129,7 +129,7 @@ public static partial class CheckValidatorsExtensions
     /// <param name="value">The number you are comparing</param>
     /// <param name="msg">Custom error message</param>
     /// <returns></returns>
-    public static Check<short?> IfNotEquals(this Check<short?> data, float value)
+    public static Check<short?> IfNotEquals(this Check<short?> data, short value)
     {
         if (data.InvalidModel()) { return data; }
         if (data.Value == value)
@@ -146,7 +146,7 @@ public static partial class CheckValidatorsExtensions
     /// <param name="value">The number you are comparing</param>
     /// <param name="msg">Custom error message</param>
     /// <returns></returns>
-    public static Check<short?> IfBetween(this Check<short?> data, double startValue, double endValue)
+    public static Check<short?> IfBetween(this Check<short?> data, short startValue, short endValue)
     {
         if (data.InvalidModel()) { return data; }
         if (data.Value > startValue && data.Value < endValue)
@@ -163,7 +163,7 @@ public static partial class CheckValidatorsExtensions
     /// <param name="value">The number you are comparing</param>
     /// <param name="msg">Custom error message</param>
     /// <returns></returns>
-    public static Check<short?> IfNotBetween(this Check<short?> data, double startValue, double endValue)
+    public static Check<short?> IfNotBetween(this Check<short?> data, short startValue, short endValue)
     {
         if (data.InvalidModel()) { return data; }
         if (data.Value <= startValue || data.Value >= endValue)
